webhook/requirements.txt
aio-pika==9.5.7
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiormq==6.9.0
aiosignal==1.4.0
attrs==25.3.0
frozenlist==1.7.0
idna==3.10
iniconfig==2.1.0
multidict==6.6.4
packaging==25.0
pamqp==3.3.0
pluggy==1.6.0
propcache==0.3.2
Pygments==2.19.2
pytest==8.4.2
pytest-aiohttp==1.1.0
pytest-asyncio==1.2.0
PyYAML==6.0.3
typing_extensions==4.15.0
yarl==1.20.1
